refactor: change directory structure for tests
