+++ conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-from test_helpers.feeder import Feeder
+from test.helpers.feeder import Feeder
 
 @pytest.fixture()
 def feeder():
