+++ test/conftest.py
@@ -2,7 +2,7 @@ import threading
 
 import pytest
 
-from test_helpers.feeder import Feeder
+from test.helpers.feeder import Feeder
 
 
 @pytest.fixture()
